Reorganize test files [ci skip]
