+++ test/blank_test.rb
@@ -1,0 +1,8 @@
+require 'test_helper'
+
+class BlankTest < Minitest::Test
+  def test_it_is_blank?
+    assert FasterPath.blank? "  "
+    assert FasterPath.blank? ""
+  end
+end
