+++ test/relative_test.rb
@@ -1,0 +1,4 @@
+require 'test_helper'
+
+class RelativeTest < Minitest::Test
+end
